feat(stats-v2): T2 extract 3 nested types to Models/Stats/
Closes F9 from v2-audit-report.md. Three types extracted from
v2/StatsView.swift into FitTracker/Models/Stats/:

- StatsPeriod.swift (47 lines) — date-range enum
- StatsFocusMetric.swift (210 lines) — metric metadata enum,
  including the `category` property added in commit 027abf0
- MetricSeriesPoint.swift (8 lines) — chart datum struct.
  Promoted from `private` to internal since it now lives in
  its own file used by the view.

v2/StatsView.swift: 899 → 673 lines (-226 net after T5/T6 add).

project.pbxproj edits (5 surgical changes, no UUID reshuffles):
- 3 new PBXBuildFile entries (SP1*/SF1*/MP1*)
- 3 new PBXFileReference entries (SP2*/SF2*/MP2*)
- New PBXGroup `Stats` (SS4*) under Models
- Models group children updated to include Stats subgroup
- 3 entries appended to the Sources build phase

Verified: xcodebuild build clean (iOS Simulator, both arm64 + x86_64).
